requirements.txt
selenium
shortuuid
docker-compose
docker
faker
robotframework
robotframework-selenium2library
robotframework-faker

